improve iv surface calc

+++ Algorithm.CSharp/Core/Indicators/IVSurfaceBin.cs
@@ -6,7 +6,7 @@ namespace QuantConnect.Algorithm.CSharp.Core.Indicators
     public class Bin
     {
         public readonly QuoteSide Side;
-        public readonly decimal Value;
+        public readonly ushort Value;
         public readonly bool IsOTM;
         public readonly DateTime Expiry;
         public readonly double Alpha;  // 1: No Smoothing, 0: No Update. First IV.
@@ -26,12 +26,12 @@ namespace QuantConnect.Algorithm.CSharp.Core.Indicators
         private DateTime PreviousTime;
         public double? Epsilon { get => IV - IVEWMA; }
 
-        private int _samples;
-        private int _smoothings;
-        public int Samples { get => _samples; }
-        public int Smoothings { get => _smoothings; }
+        private uint _samples;
+        private uint _smoothings;
+        public uint Samples { get => _samples; }
+        public uint Smoothings { get => _smoothings; }
 
-        public Bin(QuoteSide side, decimal value, DateTime expiry, bool isOTM, double alpha = 1.0, TimeSpan? samplingPeriod = null)
+        public Bin(QuoteSide side, ushort value, DateTime expiry, bool isOTM, double alpha = 1.0, TimeSpan? samplingPeriod = null)
         {
             Side = side;
             Value = value;
